+++ src/collector/CollectorWrapper.ts
@@ -32,6 +32,7 @@ import { RecordingSettingsDispatcher } from '../gravity-client/RecordingSettings
 import crossfetch from 'cross-fetch'
 import { RecordingSettings } from '../gravity-client/AbstractGravityClient'
 import ITimeoutHandler from '../timeout-handler/ITimeoutHandler'
+import ContextMenuEventListener from '../event-listeners/ContextMenuEventListener'
 
 class CollectorWrapper {
   private readonly recordingSettingsHandler = new RecordingSettingsDispatcher()
@@ -177,6 +178,9 @@ class CollectorWrapper {
     }
     if (this.isListenerEnabled(Listener.Change)) {
       eventListeners.push(new ChangeEventListener(this.userActionHandler, window, targetedEventListenerOptions))
+    }
+    if (this.isListenerEnabled(Listener.ContextMenu)) {
+      eventListeners.push(new ContextMenuEventListener(this.userActionHandler, window, targetedEventListenerOptions))
     }
     if (this.isListenerEnabled(Listener.BeforeUnload)) {
       eventListeners.push(