+++ CHANGELOG.md
@@ -19,6 +19,25 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v3.15.0...main)
+
+### Info
+
+### Added
+
+- Added new events:
+  - `contextmenu`
+
+### Changed
+
+- `data-testid` attribute is now collected by default on targeted user actions
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
 ## [3.15](https://github.com/Smartesting/gravity-data-collector/compare/v3.14.0...v3.15.0)
 
 ### Added